+++ bindings/node/native/src/tasks/tokenizer.rs
@@ -69,7 +69,7 @@ impl Task for EncodeTask<'static> {
                 Ok(js_encoding.upcast())
             }
             EncodeOutput::Batch(encodings) => {
-                let result = JsArray::new(&mut cx, encodings.len() as u32);
+                let result = JsArray::new(&mut cx, encodings.len() as u64);
                 for (i, encoding) in encodings.into_iter().enumerate() {
                     let mut js_encoding = JsEncoding::new::<_, JsEncoding, _>(&mut cx, vec![])?;
 
@@ -77,7 +77,7 @@ impl Task for EncodeTask<'static> {
                     let guard = cx.lock();
                     js_encoding.borrow_mut(&guard).encoding = Some(encoding);
 
-                    result.set(&mut cx, i as u32, js_encoding)?;
+                    result.set(&mut cx, i as u64, js_encoding)?;
                 }
                 Ok(result.upcast())
             }
@@ -86,8 +86,8 @@ impl Task for EncodeTask<'static> {
 }
 
 pub enum DecodeTask {
-    Single(Tokenizer, Vec<u32>, bool),
-    Batch(Tokenizer, Vec<Vec<u32>>, bool),
+    Single(Tokenizer, Vec<u64>, bool),
+    Batch(Tokenizer, Vec<Vec<u64>>, bool),
 }
 
 pub enum DecodeOutput {
@@ -127,10 +127,10 @@ impl Task for DecodeTask {
         match result.map_err(|e| cx.throw_error::<_, ()>(e).unwrap_err())? {
             DecodeOutput::Single(string) => Ok(cx.string(string).upcast()),
             DecodeOutput::Batch(strings) => {
-                let result = JsArray::new(&mut cx, strings.len() as u32);
+                let result = JsArray::new(&mut cx, strings.len() as u64);
                 for (i, string) in strings.into_iter().enumerate() {
                     let js_string = cx.string(string);
-                    result.set(&mut cx, i as u32, js_string)?;
+                    result.set(&mut cx, i as u64, js_string)?;
                 }
                 Ok(result.upcast())
             }
