+++ bindings/python/src/trainers.rs
@@ -179,12 +179,12 @@ impl PyBpeTrainer {
     }
 
     #[getter]
-    fn get_min_frequency(self_: PyRef<Self>) -> u32 {
+    fn get_min_frequency(self_: PyRef<Self>) -> u64 {
         getter!(self_, BpeTrainer, min_frequency)
     }
 
     #[setter]
-    fn set_min_frequency(self_: PyRef<Self>, freq: u32) {
+    fn set_min_frequency(self_: PyRef<Self>, freq: u64) {
         setter!(self_, BpeTrainer, min_frequency, freq);
     }
 
@@ -383,12 +383,12 @@ impl PyWordPieceTrainer {
     }
 
     #[getter]
-    fn get_min_frequency(self_: PyRef<Self>) -> u32 {
+    fn get_min_frequency(self_: PyRef<Self>) -> u64 {
         getter!(self_, WordPieceTrainer, min_frequency())
     }
 
     #[setter]
-    fn set_min_frequency(self_: PyRef<Self>, freq: u32) {
+    fn set_min_frequency(self_: PyRef<Self>, freq: u64) {
         setter!(self_, WordPieceTrainer, @set_min_frequency, freq);
     }
 
@@ -572,12 +572,12 @@ impl PyWordLevelTrainer {
     }
 
     #[getter]
-    fn get_min_frequency(self_: PyRef<Self>) -> u32 {
+    fn get_min_frequency(self_: PyRef<Self>) -> u64 {
         getter!(self_, WordLevelTrainer, min_frequency)
     }
 
     #[setter]
-    fn set_min_frequency(self_: PyRef<Self>, freq: u32) {
+    fn set_min_frequency(self_: PyRef<Self>, freq: u64) {
         setter!(self_, WordLevelTrainer, min_frequency, freq);
     }
 
@@ -719,12 +719,12 @@ pub struct PyUnigramTrainer {}
 #[pymethods]
 impl PyUnigramTrainer {
     #[getter]
-    fn get_vocab_size(self_: PyRef<Self>) -> u32 {
+    fn get_vocab_size(self_: PyRef<Self>) -> u64 {
         getter!(self_, UnigramTrainer, vocab_size)
     }
 
     #[setter]
-    fn set_vocab_size(self_: PyRef<Self>, vocab_size: u32) {
+    fn set_vocab_size(self_: PyRef<Self>, vocab_size: u64) {
         setter!(self_, UnigramTrainer, vocab_size, vocab_size);
     }
