+++ bindings/python/src/utils/pretokenization.rs
@@ -120,8 +120,8 @@ fn get_splits(
 
 fn to_encoding(
     pretok: &PreTokenizedString,
-    type_id: u32,
-    word_idx: Option<u32>,
+    type_id: u64,
+    word_idx: Option<u64>,
 ) -> PyResult<PyEncoding> {
     Ok(ToPyResult(
         pretok
@@ -225,7 +225,7 @@ impl PyPreTokenizedString {
     ///     An Encoding
     #[args(type_id = "0", word_idx = "None")]
     #[text_signature = "(self, type_id=0, word_idx=None)"]
-    fn to_encoding(&self, type_id: u32, word_idx: Option<u32>) -> PyResult<PyEncoding> {
+    fn to_encoding(&self, type_id: u64, word_idx: Option<u64>) -> PyResult<PyEncoding> {
         to_encoding(&self.pretok, type_id, word_idx)
     }
 
@@ -308,7 +308,7 @@ impl PyPreTokenizedStringRefMut {
     }
 
     #[args(type_id = "0", word_idx = "None")]
-    fn to_encoding(&self, type_id: u32, word_idx: Option<u32>) -> PyResult<PyEncoding> {
+    fn to_encoding(&self, type_id: u64, word_idx: Option<u64>) -> PyResult<PyEncoding> {
         self.inner
             .map(|pretok| to_encoding(pretok, type_id, word_idx))
             .ok_or_else(PyPreTokenizedStringRefMut::destroyed_error)?
