+++ tokenizers/src/normalizers/unicode.rs
@@ -46,7 +46,7 @@ fn do_nmt(normalized: &mut NormalizedString) {
     normalized
         .filter(|c| {
             !matches!(
-                c as u32,
+                c as u64,
                 0x0001..=0x0008 |
                 0x000B |
                 0x000E..=0x001F |
@@ -56,7 +56,7 @@ fn do_nmt(normalized: &mut NormalizedString) {
             )
         })
         // Other code points considered as whitespace.
-        .map(|c| match c as u32 {
+        .map(|c| match c as u64 {
             0x0009 => ' ',
             0x000A => ' ',
             0x000C => ' ',
